+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Bump postgresql from 42.4.1 to 42.4.3 in /flowman-plugins/postgresql
 * Bump loader-utils from 1.4.0 to 1.4.2
 * github-293: Fatal exceptions in parallel mapping instantiation cause deadlock
+* github-273: Support projects contained in (fat) jar files 
+* github-294: Parallel execution should not execute more targets after errors
 
 
 # Version 0.29.0 - 2022-11-08